# Conflicts:
#	release/datafiles/locale
#	release/scripts/addons
#	release/scripts/addons_contrib
#	release/scripts/startup/bl_ui/space_view3d.py
#	source/blender/editors/object/object_relations.c
#	source/blender/editors/physics/physics_pointcache.c
#	source/blender/makesrna/intern/rna_animation.c
#	source/blender/makesrna/intern/rna_fcurve.c
#	source/blender/makesrna/intern/rna_sculpt_paint.c
#	source/blender/makesrna/intern/rna_tracking.c
#	source/blender/python/intern/bpy_app_sdl.c
#	source/tools

+++ release/scripts/startup/bl_operators/wm.py
@@ -2170,7 +2170,7 @@ class WM_OT_addon_remove(Operator):
         addon_utils.disable(self.module, default_set=True)
 
         import shutil
-        if isdir:
+        if isdir and (not os.path.islink(path)):
             shutil.rmtree(path)
         else:
             os.remove(path)
